Move labels to render advert label
